Rename AllHandler to GetOpenChallengesHandler
