--- src/TicTacToe/Application/Challenge/All/AllHandler.php
@@ -1,20 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Gaming\TicTacToe\Application\Challenge\All;
-
-use Gaming\TicTacToe\Application\Model\OpenChallenges\OpenChallengesStore;
-
-final class AllHandler
-{
-    public function __construct(
-        private readonly OpenChallengesStore $openChallengesStore
-    ) {
-    }
-
-    public function __invoke(AllRequest $request): AllResponse
-    {
-        return new AllResponse($this->openChallengesStore->all($request->limit));
-    }
-}
