+++ src/TicTacToe/Application/Challenge/GetOpenChallenges/GetOpenChallengesRequest.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Gaming\TicTacToe\Application\Challenge\GetOpenChallenges;
+
+use Gaming\Common\Bus\Request;
+
+/**
+ * @implements Request<GetOpenChallengesResponse>
+ */
+final class GetOpenChallengesRequest implements Request
+{
+    public function __construct(
+        public readonly int $limit
+    ) {
+    }
+}
